requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
asyncpg==0.29.0
aiohttp==3.9.1
pytz==2023.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
pillow==10.1.0
python-dateutil==2.8.2
jdatetime==4.1.1
schedule==1.2.1
cryptography==41.0.7 